CLN reorganize the plotting system
